requirements.txt
streamlit>=1.35.0
groq>=0.9.0
pydantic>=2.6.0
python-docx>=1.1.0
openpyxl>=3.1.0
